+++ CHANGELOG.md
@@ -7,7 +7,7 @@
   * Added `robot` module that is using sessions
 * `pulseapi.robot` module
   * Added `RobotPulse.zg_on()` and `RobotPulse.zg_off()` methods that control freedrive mode.
-  * Added `RobotPulse.stop()`
+  * Added `RobotPulse.stop()`, `RobotPulse.bind_stop` and `RobotPulse.unbind_stop()`
 
 
 ## 1.7.0: